Make get_channel available as generic helper
