+++ homeassistant/components/sentry/__init__.py
@@ -15,7 +15,7 @@ from homeassistant.const import (
     EVENT_HOMEASSISTANT_STARTED,
     __version__ as current_version,
 )
-from homeassistant.core import HomeAssistant
+from homeassistant.core import HomeAssistant, get_channel
 from homeassistant.helpers import config_validation as cv, entity_platform, instance_id
 from homeassistant.helpers.event import async_call_later
 from homeassistant.helpers.system_info import async_get_system_info
@@ -108,17 +108,6 @@ async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     hass.bus.async_listen_once(EVENT_HOMEASSISTANT_STARTED, update_system_info)
 
     return True
-
-
-def get_channel(version: str) -> str:
-    """Find channel based on version number."""
-    if "dev0" in version:
-        return "dev"
-    if "dev" in version:
-        return "nightly"
-    if "b" in version:
-        return "beta"
-    return "stable"
 
 
 def process_before_send(
