+++ tests/components/sentry/test_init.py
@@ -4,7 +4,7 @@ from unittest.mock import Mock, patch
 
 import pytest
 
-from homeassistant.components.sentry import get_channel, process_before_send
+from homeassistant.components.sentry import process_before_send
 from homeassistant.components.sentry.const import (
     CONF_DSN,
     CONF_ENVIRONMENT,
@@ -101,20 +101,6 @@ async def test_setup_entry_with_tracing(hass: HomeAssistant) -> None:
         "traces_sample_rate",
     }
     assert call_args["traces_sample_rate"] == 0.5
-
-
-@pytest.mark.parametrize(
-    ("version", "channel"),
-    [
-        ("0.115.0.dev20200815", "nightly"),
-        ("0.115.0", "stable"),
-        ("0.115.0b4", "beta"),
-        ("0.115.0dev0", "dev"),
-    ],
-)
-async def test_get_channel(version: str, channel: str) -> None:
-    """Test if channel detection works from Home Assistant version number."""
-    assert get_channel(version) == channel
 
 
 async def test_process_before_send(hass: HomeAssistant) -> None:
