+++ tests/test_core.py
@@ -43,6 +43,7 @@ from homeassistant.core import (
     State,
     SupportsResponse,
     callback,
+    get_channel,
 )
 from homeassistant.exceptions import (
     HomeAssistantError,
@@ -2481,3 +2482,20 @@ async def test_validate_state(hass: HomeAssistant) -> None:
     assert ha.validate_state("test") == "test"
     with pytest.raises(InvalidStateError):
         ha.validate_state("t" * 256)
+
+
+@pytest.mark.parametrize(
+    ("version", "channel"),
+    [
+        ("0.115.0.dev20200815", "nightly"),
+        ("0.115.0", "stable"),
+        ("0.115.0b4", "beta"),
+        ("0.115.0dev0", "dev"),
+    ],
+)
+async def test_get_channel(version: str, channel: str) -> None:
+    """Test if channel detection works from Home Assistant version number."""
+    assert get_channel(version) == channel
+
+    with patch("homeassistant.core.__version__", f"{version}"):
+        assert get_channel() == channel
